+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 更新日志 / Change Log
 
 1. 修复已知BUG / Fix known bugs
+2. 修复视频采集功能 / Fix video capture function